Vardiya listeleme ve ekleme tamam

+++ MSSWebUI/Controllers/ShowCaseController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MSSWebUI.Models.DTO;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -26,9 +27,30 @@ namespace MSSWebUI.Controllers
 
         }
 
+        public bool SessionKontrol()
+        {
+            try
+            {
+                _employee = JsonConvert.DeserializeObject<Employee>(HttpContext.Session.GetString("SessionUser"));
+
+
+                bool result = (_employee == null) ? false : true;
+                return result;
+            }
+            catch (Exception)
+            {
+
+                return false;
+            }
+        }
+
         public IActionResult Index()
         {
-
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
 
             AddShowCaseShopDTO addShowCaseShopDTO = new AddShowCaseShopDTO();
             addShowCaseShopDTO.ShowCaseDetails= _showCaseService.GetShowCaseDetails();
@@ -38,7 +60,11 @@ namespace MSSWebUI.Controllers
         }
         public IActionResult AddShowCase()
         {
-
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
 
             AddShowCaseShopDTO addShowCaseShopDTO = new AddShowCaseShopDTO();
             addShowCaseShopDTO.ShowCaseDetails = _showCaseService.GetShowCaseDetails();
@@ -49,6 +75,11 @@ namespace MSSWebUI.Controllers
         
         public IActionResult AddImage(int id)
         {
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
             AddShowCaseShopDTO addShowCaseShopDTO = new AddShowCaseShopDTO();
             addShowCaseShopDTO.Shop= _shopService.GetByShopId(id);            
             return View(addShowCaseShopDTO);
@@ -59,7 +90,11 @@ namespace MSSWebUI.Controllers
         public async Task<IActionResult> Upload(int id, IFormFile file)
 
         {
-            
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
 
             var FileDic = "Upload\\ShowCase\\"+id+"\\"+DateTime.Now.ToShortDateString();
 
@@ -86,7 +121,7 @@ namespace MSSWebUI.Controllers
                 _showCaseService.Add(showCase);
 
                 file.CopyTo(fs);
-
+                
 
             }
 
@@ -97,9 +132,25 @@ namespace MSSWebUI.Controllers
         [HttpPost]
         public IActionResult DeleteShowCase(ShowCase showCase)
         {
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
+
+
+            string FilePath = Path.Combine(_environment.WebRootPath, showCase.Url);
 
             try
             {
+               
+                if (System.IO.File.Exists(FilePath))
+                {
+                    // If file found, delete it    
+                    System.IO.File.Delete(FilePath);
+                    
+                }
+
                 _showCaseService.Delete(showCase);
             }
             catch (Exception)
@@ -112,6 +163,13 @@ namespace MSSWebUI.Controllers
         [HttpPost]
         public IActionResult UpdateShowCase(AddShowCaseShopDTO addShowCaseShopDTO)
         {
+
+            if (!SessionKontrol())
+            {
+                HttpContext.Session.Clear();
+                return RedirectToAction("Index", "Home");
+            }
+
             addShowCaseShopDTO.ShowCase.ShopId = addShowCaseShopDTO.Shop.ShopId;
